+++ demo/HuggingFace/T5/onnxrt.py
@@ -221,7 +221,7 @@ class T5ONNXRT(OnnxRTCommand):
 
         return NetworkResult(
             input=inference_input,
-            output_tensor=encoder_last_hidden_state,
+            output_tensor=decoder_output,
             semantic_output=semantic_outputs,
             median_runtime=runtime,
             models=models,